Resume Critiquer/pyproject.toml
[project]
name = "resume-critiquer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain-google-genai>=2.0.10",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
]
